requirements.txt
setuptools~=58.0.4
pyyaml~=6.0
python-binance~=1.0.15
pickle5~=0.0.11